+++ packages/create-docgen/templates/starter/src/routes/+layout.server.ts
@@ -1,7 +1,0 @@
-import { makeRootLayoutLoad } from '@iammaxim/docgen/server';
-import { registry } from '$lib/docgen-registry';
-import type { LayoutServerLoad } from './$types';
-
-export const prerender = true;
-
-export const load: LayoutServerLoad = makeRootLayoutLoad(registry);